requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.10.0
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.3
coqpit-config==0.2.1
coqui-tts==0.27.1
coqui-tts-trainer==0.3.1
cycler==0.12.1
Cython==3.1.3
dateparser==1.1.8
decorator==5.2.1
docopt==0.6.2
dora_search==0.1.12
einops==0.8.1
encodec==0.1.1
face-alignment==1.4.1
fastapi==0.116.2
ffmpeg-python==0.2.0
ffmpy==0.6.1
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
gradio==5.45.0
gradio_client==1.13.0
groovy==0.1.2
grpcio==1.74.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.5
idna==3.10
imageio==2.37.0
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
julius==0.2.7
kiwisolver==1.4.9
lameenc==1.8.1
lazy_loader==0.4
librosa==0.10.2.post1
lipsync==0.1.0
llvmlite==0.44.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
monotonic-alignment-search==0.2.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
networkx==3.5
num2words==0.5.14
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.12.0.88
openunmix==1.3.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.4
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
retrying==1.4.2
rich==14.1.0
ruff==0.13.0
sacremoses==0.1.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
srt==3.5.3
starlette==0.48.0
submitit==1.5.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.9.9
tokenizers==0.22.0
tomlkit==0.13.3
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
transformers==4.56.1
treetable==0.2.6
triton==3.4.0
typeguard==4.4.4
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1